remove useless stuff

+++ BaselineGAN/FusedOperators.py
@@ -14,7 +14,7 @@ class BiasedActivationReference(nn.Module):
         self.Bias.data.zero_()
         
     def forward(self, x):
-        y = x + self.Bias.view(1, -1, 1, 1) if len(x.shape) > 2 else x + self.Bias.view(1, -1)
+        y = x + self.Bias.to(x.dtype).view(1, -1, 1, 1) if len(x.shape) > 2 else x + self.Bias.to(x.dtype).view(1, -1)
         return BiasedActivationReference.Function(y)
 
 class BiasedActivationCUDA(nn.Module):
